Calc functions to own module * shower calling Python `electr` * working through `electr` bugs
